competitions/stanford-rna-3d-folding/requirements.txt
# Deep Learning
torch>=2.0.0
pytorch-lightning>=2.0.0

# Molecular Biology
biopython>=1.81
biotite>=0.37.0
rdkit>=2023.3.1

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
scikit-learn>=1.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
py3Dmol>=2.0.0
plotly>=5.15.0

# Data Processing
h5py>=3.8.0
networkx>=3.0
prody>=2.4.0

# Machine Learning Tools
wandb>=0.15.0  # for experiment tracking
hydra-core>=1.3.0  # for configuration management
pytest>=7.3.1  # for testing
black>=23.3.0  # for code formatting
flake8>=6.0.0  # for code linting

# Kaggle
kaggle>=1.5.13 